requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
pandas==2.1.3
python-dotenv==1.0.0
Werkzeug==3.0.1
reportlab==4.4.7
